Refactor so that the Trait gets the name BorrowedPix
I think this is clearer.
